Move new(), capacity(), and with_capacity() funcs to their own file
